separate php 7.x expectations
